Re-organise function includes
